fix: refactoring the api keys section and add documentation link.
